requirement.txt
appnope             ==0.1.0    
asn1crypto          ==0.24.0   
astroid             ==2.0.2    
backcall            ==0.1.0    
certifi             ==2018.4.16
cffi                ==1.11.5   
chardet             ==3.0.4    
coreapi             ==2.3.3    
coreschema          ==0.0.4    
cryptography        ==2.3      
decorator           ==4.3.0    
Django              ==2.1      
django-filter       ==2.0.0    
djangorestframework ==3.8.2    
idna                ==2.7      
ipython             ==6.4.0    
ipython-genutils    ==0.2.0    
isort               ==4.3.4    
itypes              ==1.1.0    
jedi                ==0.12.1   
Jinja2              ==2.10     
lazy-object-proxy   ==1.3.1    
Markdown            ==2.6.11   
MarkupSafe          ==1.0      
mccabe              ==0.6.1    
mysqlclient         ==1.3.13   
parso               ==0.3.1    
pexpect             ==4.6.0    
pickleshare         ==0.7.4    
pip                 ==18.0     
prompt-toolkit      ==1.0.15   
ptyprocess          ==0.6.0    
pycparser           ==2.18     
Pygments            ==2.2.0    
pylint              ==2.1.1    
pytz                ==2018.5   
requests            ==2.19.1   
setuptools          ==39.0.1   
simplegeneric       ==0.8.1    
six                 ==1.11.0   
traitlets           ==4.3.2    
typed-ast           ==1.1.0    
typing              ==3.6.4    
uritemplate         ==3.0.0    
urllib3             ==1.23     
wcwidth             ==0.1.7    
wrapt               ==1.10.11  
==